feat(user): Update functions logic
Edit addRecipeToUserFavourite() and removeRecipeFromUserFavourite() to first check the output of a new isRecipeInUserFavorite() function.
This function is used to check if the recipe is already present (when trying to add) or if it is not present (when trying to remove) with related error messages that will be prompted to the user.