feat!(recipeCommand): Rework buttons logic
Edit image visualization, don't show any image if it is not present (this has been done since the previous solution was too imposing beign the logo orange).
Remove icon from footer since the image will be already present with the message author.
Make the buttons have a dynamic id; They now have their id linked to the recipe they are attached to. This way we are able to resolve which recipe was selected when an "Add" button is pressed.
Remove button and confirmation logic; This will be implemeted in the file handling buttons.

BREAKING CHANGE: Buttons now are linked to the recipe they are attached to and when pressed they must be resolved accordingly.