feat!(buttonHandler): Add new button handling system
Add a new way to handle a "permanent component" button. This are buttons which can be pressed anytime after they have been created.

button naming is set as follows: "button label":"button value"
(Eg. add_favorite_recipe:64f1b2e6d71f6dbd86419093)

The first part - button label - is the one that is checked to know what type of button was pressed, the second part - button value - (after the ":") has the unique ID identifing the message to which the button is attached to, in this case the recipeId, and is used to find the corresponding recipe.

This method does not need "confirmation" since the buttons are not time dependant and the interaction is not wating for them to be interacted with but is called when the button is pressed.

BREAKING CHANGE: Some buttons are now "permanent component buttons" and must be handled for and in this file.