feat(favoriteCommand): Edit list visualization
Rename "favourites" to "favorite" for cohesion with rest of the code.
Edit recipe descriptions to be at max 100 characters long to not have a wall of text; If the description would have been longer it will end with a whole word and "..." (Eg. "with a variety of...").
Add a thumbnail showing the recipe image to better visualize the recipes.