feat!(user.model): Update user schema
Update favorite recipes list to include a link-date pair instead of just a list of links.
"date" will be equal to the moment the recipe was added to favorites for stats (tbd) and ux.
Each recipe in the favorite list could then include a "Added on 23/10/2023" note.

BREAKING CHANGE: User shema changed to include a date for each url entry.