refactor(favoriteCommand): Edit favorites list visualization
Edit visualization to condense informations, now the list will be shown in one embed and without pagination since "forward" and "backward" buttons have issues. This visualization is currently conditional (line 141) and is currently the default one, it could later be edited to be set by the user (User could have a "setting" tab to choose what list type they want).