feat(listCommand): Edit error response when no category is found
Add a "Contact support" button to the error message when no category is found. Error message is now linked to the language. This method will likely be added to all (or most) error messages to directly prompt the user to contact support and give some context of what is happening other than "The Application did not respond" or similar.
Comment out parts of codes related to the "/list country" command as it does not seem to be working right now.